+++ lib/recurly/add_on.rb
@@ -6,9 +6,11 @@ module Recurly
     define_attribute_methods %w(
       add_on_code
       name
+      accounting_code
       default_quantity
       unit_amount_in_cents
       display_quantity_on_hosted_page
+      tax_code
       created_at
     )
     alias to_param add_on_code