+++ lib/recurly/adjustment.rb
@@ -29,11 +29,13 @@ module Recurly
       total_in_cents
       currency
       tax_exempt
+      tax_code
       tax_details
       product_code
       start_date
       end_date
       created_at
+      quantity_remaining
     )
     alias to_param uuid