+++ spec/recurly/subscription_spec.rb
@@ -26,10 +26,16 @@ describe Subscription do
                                 po_number
                                 tax_in_cents
                                 tax_type
+                                tax_region
                                 tax_rate
                                 total_billing_cycles
                                 remaining_billing_cycles
-                                bulk}
+                                bulk
+                                terms_and_conditions
+                                customer_notes
+                                address
+                                vat_reverse_charge_notes
+                              }
 
         subject.attribute_names.sort.must_equal expected_attributes.sort
       end
@@ -40,6 +46,8 @@ describe Subscription do
       {
         plan_code: 'gold',
         currency: 'EUR',
+        terms_and_conditions: 'Some Terms and Conditions',
+        customer_notes: 'Some Customer Notes',
         account: {
           account_code: '1',
           email: 'verena@example.com',
@@ -277,6 +285,21 @@ describe Subscription do
       subscription.plan.plan_code.must_equal 'plan_code'
     end
 
+    it 'should be able to derive and parse address' do
+      stub_api_request :post, 'subscriptions/preview', 'subscriptions/preview-200-new'
+
+      subscription = Subscription.preview(
+        plan_code: 'plan_code',
+        currency: 'USD',
+        account: {
+          account_code: 'account_code',
+        }
+      )
+
+      subscription.address.must_be_instance_of Address
+      subscription.address.country.must_equal 'US'
+    end
+
     it 'previews subscription changes' do
       stub_api_request :get, 'subscriptions/abcdef1234567890', 'subscriptions/show-200-noinvoice'
       stub_api_request :post, 'subscriptions/abcdef1234567890/preview', 'subscriptions/preview-200-change'
@@ -287,6 +310,26 @@ describe Subscription do
 
       subscription.cost_in_cents.must_equal subscription.unit_amount_in_cents * 5
       subscription.invoice.must_be_instance_of Invoice
+    end
+  end
+  describe 'notes' do
+    it 'previews new subscriptions' do
+      stub_api_request :get, 'subscriptions/abcdef1234567890', 'subscriptions/show-200'
+      stub_api_request :put, 'subscriptions/abcdef1234567890/notes', 'subscriptions/notes-200-change'
+
+      subscription = Subscription.find 'abcdef1234567890'
+
+      notes = {
+        customer_notes: 'Some New Customer Notes',
+        terms_and_conditions: 'Some New Terms and Conditions',
+        vat_reverse_charge_notes: 'Some New Vat Reverse Charge Notes'
+      }
+
+      subscription.update_notes(notes)
+
+      subscription.customer_notes.must_equal notes[:customer_notes]
+      subscription.terms_and_conditions.must_equal notes[:terms_and_conditions]
+      subscription.vat_reverse_charge_notes.must_equal notes[:vat_reverse_charge_notes]
     end
   end
 end